backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
numpy
scikit-learn
Pillow
python-multipart